Run accessibility tests on all variations
